templating | ✨ feat: Add template support for commit messages
This commit adds support for using templates in commit messages. The `processTemplate` function replaces `{COMMIT_MESSAGE}` with the actual commit message and optionally replaces `templating` with the current branch name if present in the template. If the template does not include `{COMMIT_MESSAGE}`, a warning is displayed. The `generateSingleCommit` and `generateListCommits` functions utilize the template processing logic to generate the final commit messages. The proposed commit message is displayed, either with or without the template, depending on whether a template is provided. The `makeCommit` function is called to commit the final message.